Adjust tests to work with context provider
